+++ src/components/transactions_list/components/mobile/index.tsx
@@ -1,11 +1,15 @@
 import React from 'react';
 import dayjs from '@utils/dayjs';
+import Link from 'next/link';
 import useTranslation from 'next-translate/useTranslation';
 import { getMiddleEllipsis } from '@utils/get_middle_ellipsis';
 import {
   Typography, Divider,
 } from '@material-ui/core';
 import { Result } from '@components';
+import {
+  TRANSACTION_DETAILS, ACCOUNT_DETAILS,
+} from '@utils/go_to_page';
 import { useStyles } from './styles';
 import { Shard } from '..';
 
@@ -14,16 +18,34 @@ const Mobile: React.FC<{items: TransactionType[]} & ComponentDefault> = (props) 
   const classes = useStyles();
   const formattedItems = props.items.map((x) => {
     return ({
-      hash: getMiddleEllipsis(x.hash, {
-        beginning: 13, ending: 15,
-      }),
+      hash: (
+        <Link href={TRANSACTION_DETAILS(x.hash)} passHref>
+          <Typography variant="body1" className="value" component="a">
+            {getMiddleEllipsis(x.hash, {
+              beginning: 13, ending: 15,
+            })}
+          </Typography>
+        </Link>
+      ),
       shard: <Shard to={x.toShard} from={x.fromShard} />,
-      from: getMiddleEllipsis(x.from, {
-        beginning: 13, ending: 15,
-      }),
-      to: getMiddleEllipsis(x.to, {
-        beginning: 13, ending: 15,
-      }),
+      from: (
+        <Link href={ACCOUNT_DETAILS(x.from)} passHref>
+          <Typography variant="body1" className="value" component="a">
+            {getMiddleEllipsis(x.from, {
+              beginning: 13, ending: 15,
+            })}
+          </Typography>
+        </Link>
+      ),
+      to: (
+        <Link href={ACCOUNT_DETAILS(x.to)} passHref>
+          <Typography variant="body1" className="value" component="a">
+            {getMiddleEllipsis(x.to, {
+              beginning: 13, ending: 15,
+            })}
+          </Typography>
+        </Link>
+      ),
       status: (
         <Result status={x.status} />
       ),