+++ src/components/transactions_list/components/shard/index.tsx
@@ -1,14 +1,23 @@
 import React from 'react';
+import useTranslation from 'next-translate/useTranslation';
 import { ArrowForward } from '@material-ui/icons';
+import { getShardDisplay } from '@utils/get_shard_display';
 import { useStyles } from './styles';
 
 const Shard: React.FC<{to: number, from: number} & ComponentDefault> = (props) => {
+  const { t } = useTranslation('blocks');
   const classes = useStyles();
+  const from = getShardDisplay(props.from);
+  const to = getShardDisplay(props.to);
   return (
     <div className={classes.root}>
-      {props.from}
+      {t(from.key, {
+        num: from.num,
+      })}
       <ArrowForward className={classes.icon} />
-      {props.to}
+      {t(to.key, {
+        num: to.num,
+      })}
     </div>
   );
 };