+++ src/screens/transaction_details/components/operations/components/desktop/utils.ts
@@ -1,0 +1,22 @@
+export const columns:{
+  key: string;
+  align?: 'left' | 'center' | 'right' | 'justify' | 'inherit';
+  width: number;
+}[] = [
+  {
+    key: 'action',
+    width: 20,
+  },
+  {
+    key: 'sender',
+    width: 30,
+  },
+  {
+    key: 'receiver',
+    width: 30,
+  },
+  {
+    key: 'identifier',
+    width: 20,
+  },
+];